Add API tests
Align tests file structure with that of src

Add API test

Add api test
