+++ tests/modules/util.js
@@ -3,49 +3,73 @@
  * All rights reserved. Use of this material is subject to license.
  */
 
-const cli = require('../../src/functions/cli');
+const { expect } = require('chai');
+const sinon = require('sinon');
+const prompt = require('../../src/modules/prompt');
 const switches = require('../../src/modules/switches');
+const util = require('../../src/modules/util');
 
-const ctx = {};
-
-const waitAsync = async ms => new Promise(resolve => setTimeout(resolve, ms));
-
-const createProject = async () => {
-  const payload = JSON.stringify({ name: 'Test project' });
-  const res = await cli(`projects create ${payload}`);
-
-  return res.data;
-};
-
-const createApplication = async (projectId) => {
-  const payload = JSON.stringify({ name: 'Test app', socialNetworks: {} });
-  const res = await cli(`projects ${projectId} applications create ${payload}`);
-
-  return res.data;
-};
-
-const createAppUser = async (name) => {
-  let payload = JSON.stringify({
-    firstName: 'test',
-    lastName: 'user',
-    email: `${name}@example.com`,
-    password: 'password',
+describe('util', () => {
+  afterEach(() => {
+    switches.BUILD = false;
+    sinon.resetBehavior();
   });
-  let res = await cli(`app-users create ${payload} --api-key ${ctx.application.appApiKey}`);
-  switches.API_KEY = false;
 
-  // Validate Application User
-  payload = JSON.stringify({ activationCode: res.data.activationCode });
-  const argStr = `app-users ${res.data.evrythngUser} validate ${payload} --api-key ${ctx.application.appApiKey}`;
-  res = await cli(argStr);
-  switches.API_KEY = false;
-  return res.data;
-};
+  it('should know an ID is 24 characters long', () => {
+    const id = 'UKAVpbnsVDPa9Kaaam7a5tdp';
 
-module.exports = {
-  ctx,
-  waitAsync,
-  createAppUser,
-  createProject,
-  createApplication,
-};
+    const isId = () => util.isId(id);
+    expect(isId()).to.equal(true);
+  });
+
+  it('should pretty format a JSON string', () => {
+    const json = [{key: 'value'}];
+    const expected = '[\n  {\n    "key": "value"\n  }\n]';
+
+    const pretty = () => util.pretty(json);
+    expect(pretty()).to.equal(expected);
+  });
+
+  it('should not throw error for printListSummary()', () => {
+    const list = [
+      { id: 'UKAVpbnsVDPa9Kaaam7a5tdp', name: 'thng1' },
+      { id: 'UKAVpbnsVDPa9Kaaam7a5td2', name: 'thng2' }
+    ];
+
+    const printListSummary = () => util.printListSummary(list);
+    expect(printListSummary).to.not.throw();
+  });
+
+  it('should not throw error for printSimple()', () => {
+    const data = [
+      { id: 'UKAVpbnsVDPa9Kaaam7a5tdp', name: 'thng1' },
+      { id: 'UKAVpbnsVDPa9Kaaam7a5td2', name: 'thng2' }
+    ];
+
+    const printSimple = () => util.printSimple(data);
+    expect(printSimple).to.not.throw();
+  });
+
+  it('should not throw and error for requireKey', () => {
+    const key = '12345687123456812345678123456781234567812345678123456871234568712345678123465781';
+    switches.API_KEY = key;
+
+    const requireKey = () => util.requireKey('Application');
+    expect(requireKey).to.not.throw();
+  });
+
+  it('should build a correct thng payload using the user prompts', async () => {
+    switches.BUILD = true;
+    sinon.stub(prompt, 'getValue')
+      .onCall(0).returns('TestThng')
+      .returns('');
+
+    const payload = await util.getPayload('ThngDocument');
+    expect(payload).to.have.property('name', 'TestThng');
+  });
+
+  it('should build the correct thng payload using JSON', async () => {
+    const payload = await util.getPayload('ThngDocument', '{"name":"TestThng"}');
+    expect(payload).to.have.property('name', 'TestThng');
+  });
+});
